# api-gateway/src/docs/openapi3_0.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Supermarket API",
    "version": "1.0.0",
    "description": "A Node.js Express API for managing supermarket-related services including authentication, users, products, and purchases. The system is designed with scalability and maintainability in mind, separating concerns into independent services that follow the  Model-View-Controller(MVC) design pattern and communicate through an API Gateway."
  },
  "servers": [
    {
      "url": "http://localhost:3000/supermarket/",
      "description": "Base URL for the supermarket API"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieve all users",
        "operationId": "getUsers",
        "description": "Fetches a list of all users from the database, ordered by last update time in descending order",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "description": "Total number of users",
                      "example": 8
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user",
        "operationId": "createUser",
        "description": "Creates a new user with the provided details after validating input data",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User Created successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpressValidatorError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieve a user by ID",
        "operationId": "getUserById",
        "description": "Fetches a specific user from the database using their unique identifier (UUID)",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the user (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3ba4dea-e55f-423d-b863-3eb5bdb79fdc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User founded successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidIdError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update a user by ID",
        "operationId": "updateUser",
        "description": "Updates an existing user's details based on their unique identifier (UUID). All fields are optional.",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the user (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3ba4dea-e55f-423d-b863-3eb5bdb79fdc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or Invalid Id",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ExpressValidatorError"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidIdError"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": null,
        "operationId": "deleteUser",
        "description": null,
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the user (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3ba4dea-e55f-423d-b863-3eb5bdb79fdc"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully (no content returned)"
          },
          "400": {
            "description": "Invalid ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidIdError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Cannot delete Super Admin user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "impossible to delete de Super Admin user"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/usernameOrEmail/{value}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Retrieve a user by username or email",
        "operationId": "getUserByUsernameOrEmail",
        "description": "Fetches a user from the database using either their username or email address",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "description": "The username or email address of the user",
            "required": true,
            "schema": {
              "type": "string",
              "example": "johndoe | john.doe@example.com"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User founded successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Username or email not provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Username or Email no provider"
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "BadRequest"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/resetPassword/{id}": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Reset a user's password",
        "operationId": "resetPassword",
        "description": "Resets a user's password by verifying their current password and updating it with a new one",
        "security": [
          {
            "AccessToken": []
          },
          {
            "RefreshToken": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique identifier of the user (UUID format).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "b3ba4dea-e55f-423d-b863-3eb5bdb79fdc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password reset successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpressValidatorError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Impossible to reset password, user not found"
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "UserNotFound"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Current password incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Passwords do not match."
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "InvalidPass"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the user.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Michael Kenneth Williams"
          },
          "username": {
            "type": "string",
            "description": "Username chosen by the user for login or identification purposes.",
            "example": "mkwilliams"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user, used for communication.",
            "example": "michael.williams@x.super.com\""
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user, defining their permissions and      access level.",
            "enum": [
              "admin",
              "manager",
              "supervisor",
              "user"
            ],
            "example": "USER"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was created.",
            "example": "2025-03-15T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp indicating when the user account was last updated.",
            "example": "2025-03-25T11:56:00Z"
          }
        }
      },
      "UserRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the user.",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Michael Kenneth Williams"
          },
          "username": {
            "type": "string",
            "description": "Username chosen by the user for login or identification purposes.",
            "example": "mkwilliams"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user, used for communication.",
            "example": "michael.williams@x.super.com\""
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user, defining their permissions and access level.",
            "enum": [
              "admin",
              "manager",
              "supervisor",
              "user"
            ],
            "example": "USER"
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "Michael Kenneth Williams"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the user, used for communication.",
            "example": "michael.williams@x.super.com\""
          },
          "role": {
            "type": "string",
            "description": "Role assigned to the user, defining their permissions and access level.",
            "enum": [
              "admin",
              "manager",
              "supervisor",
              "user"
            ],
            "example": "supervisor"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "The user's current password for verification"
          },
          "newPassword": {
            "type": "string",
            "description": "New password, must have at least one number, letter, and special character and with a length between 6 and 12 characters."
          },
          "confirm": {
            "type": "string",
            "description": "Confirmation of the new password, must match newPassword"
          }
        },
        "required": [
          "password",
          "newPassword",
          "confirm"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ExpressValidatorError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates whether the request was successful or not.",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "A general message describing the result of the request.",
            "example": "validation failed"
          },
          "errors": {
            "type": "array",
            "description": "A list of validation errors encountered during the request.",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field with validation error",
                  "example": "stock"
                },
                "message": {
                  "type": "string",
                  "description": "Descriptive error message",
                  "example": "Expected number, received string"
                }
              }
            }
          }
        }
      },
...
      "InvalidIdError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The provided ID does not have a valid format"
          },
          "errorCode": {
            "type": "string",
            "example": "InvalidId"
          }
        }
      },
...
      "UserNotFound": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "User not found"
          },
          "user": {
            "type": "object",
            "example": null
          }
        }
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "authorization",
        "description": "Access token passed in the Authorization header"
      },
      "RefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Refresh token passed as refresh_token=bearer<JWT> in the Cookie header"
      }
    }
  }
}